FileUtilities Testing
